add charageable spell functionallty, now fix it

+++ src/main/java/fn10/desktopClicker/game/spells/ICharageableSpell.java
@@ -4,7 +4,7 @@ import fn10.desktopClicker.game.SavedGame;
 
 public interface ICharageableSpell extends IBaseSpell {
 
-    void cast(SavedGame game, int manaSpent);
+    void cast(SavedGame game, float manaSpent);
 
     float getManaRequirment(SavedGame game, float timeHeld);
